14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,9 +11,9 @@ int main(void)
 
 {
 
-	int i, n;
-	int t1 = 1, t2 = 2;
-	int nextTerm = t1 + t2;
+	long int i, n;
+	long int t1 = 1, t2 = 2;
+	long int nextTerm = t1 + t2;
 
 	n = 50;
 	printf("%d, %d, ", t1, t2);
